Refactor (01/12/22)
Separate GUI update methods from InputProcessor into CalculatorGuiUpdater. Add delegates to InputProcessor to signal to CalculatorGuiUpdater to update various aspects of the GUI.

Update Input Processor's "ClearInput()" method. Instead of determining the state of the calculator's input based on GUI elements, determines whether it should clear the operand or current expression based on a new boolean, "ClearedOnce". If this boolean is true, it clears the expression, otherwise, it clears the operand.

Update InputProcessorTests to reflect changes to InputProcessor's "ClearInput()".